requirements.txt
droidrun[google]
google-genai
uiautomator2
python-dotenv